Address review: per-row-group passthrough values, and recover after a reply
Four real defects from CodeRabbit on #192, plus three smaller corrections.

**Passthrough values were bound to the wrong row group.** `PassthroughColumn` carried
the decoded `values` alongside its accumulating buffer, resolved once against the first
table. A tile spanning two row groups then scanned the second table with `rowIndex`
restarting at zero while reading the first group's values — every point got exactly one
value, so neither the lockstep discipline nor the output-length checks could see it.
The fix is structural rather than a patch: the column no longer holds values at all, and
the scan re-reads them from the table in front of it. Regression test covers two row
groups.

**Bool columns came back as NaN.** `DataType.isBool` was accepted, but
`numericColumnValues` keeps only `typeof value === 'number'` on its boxed path, so an
`overlaps_nucleus` written as Bool produced a full-length array of NaN — the same
silently-aligned, silently-meaningless result the type guard was added to prevent.
Converted explicitly, nulls kept distinct.

**Int32 above 2^24 was rounded.** Rejecting 64-bit integers was not enough: a
`Float32Array` rounds 16,777,217. Passthrough columns now travel in a `Float64Array`,
the one lane that carries every accepted type — int32, float32, float64, bool —
exactly.

**Crash recovery never ran for a worker that had answered.** The `workerHasAnswered`
branch returned before the `workerLoaded` check, and every successful response sets it.
Since metadata and catalog requests usually succeed before a refused range panics, this
was the common case: the dead worker stayed installed with `isParquetWorkerEnabled()`
true, so later requests waited out the silence watchdog. It also made the restart budget
unreachable, contradicting this file's own claim that a long session keeps recovering.
A loaded worker is now replaced whether or not it has replied.

Also: bound `ParquetFile.fromUrl()` and `file.stream()` with the same budget as the
reads, since the panic can land while opening; correct a watchdog label that reported
the feature-catalog scan as an in-bounds stream; make the writer reject an unrecognised
encoding policy rather than silently writing pyarrow defaults, and record `None` plus
the policy name instead of an empty plan that reads as "no column uses a dictionary" —
the opposite of what that policy does; fix two stale cross-references in the plan doc.

+++ packages/core/src/workers/parquetWorkerProtocol.ts
@@ -246,8 +246,12 @@ export type ParquetWorkerColumnarResult = {
    * in lockstep with `xs`/`ys`. A requested name is absent here when the column is
    * missing, non-numeric, or 64-bit — the worker warns rather than returning values
    * the caller cannot trust.
+   *
+   * `Float64Array` because it is the one lane that carries every accepted source type
+   * exactly: int32, float32, float64 and bool all round-trip, where `Float32Array`
+   * would quietly round an Int32 above 2^24.
    */
-  columns?: Record<string, Float32Array>;
+  columns?: Record<string, Float64Array>;
 };
 
 export type ParquetWorkerScanResult = Omit<ParquetWorkerColumnarResult, 'kind'> & {
